refactor: extract shared admin REST utilities into _shared/adminRest
- Extract adminFetch, validateUuid, createRestPath, getAuthorizedAdmin,
  fetchGrantedPermissions, fetchGrantedPermissionKeys into shared module
- Update permissions route to import from shared module
- Update delegates route to import from shared module
- Eliminates duplicated auth/REST boilerplate across API routes
